source/requirements-build-and-test.txt
black
boto3
click
coverage
crhelper==2.0.6
jinja2
pytest
pytest-cov
pytest-mock
pytest-env
moto==1.3.15.dev882
requests-mock
